requirements.txt
Arpeggio==1.9.2
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
dill==0.3.1.1
grpcio==1.29.0
hvac==0.10.3
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
parver==0.3.0
protobuf==3.12.2
pulumi==2.3.0
pulumi-vault==2.2.0
pulumi-vsphere==2.3.4
pyvmomi==7.0
PyYAML==5.3.1
requests==2.23.0
semver==2.10.1
six==1.15.0
urllib3==1.25.9
